// libs/enums/board-article.enum.ts
export enum BoardArticleCategory {
	FREE = 'FREE',
	RECOMMEND = 'RECOMMEND',
	NEWS = 'NEWS',
	HUMOR = 'HUMOR',
	STYLE_TALK = 'STYLE_TALK',
	FIT_CHECK = 'FIT_CHECK',
	MARKET_NEWS = 'MARKET_NEWS',
	SELLER_NOTES = 'SELLER_NOTES',
}

export enum BoardArticleStatus {
	ACTIVE = 'ACTIVE',
	DELETE = 'DELETE',
}